requirements.txt
pocketflow>=0.1.0
openai>=1.4.0
anthropic>=0.7.0
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.45.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0 